+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Changes
 
+* Removed `cider-clojure-cli-parameters` due to clojure-cli jack-in changes
 * Bump the injected `cider-nrepl` to 0.25.6. This should fix a compatibility issue with Java 15 and fetching fresh ClojureDocs data.
 
 ## 1.0.0 (2020-28-12)